Gas dropoff slider conversion rate fix
Previously the gas dropoff slider was setting the native gas amount
using the conversion rate from CoinGecko. However the conversion rate
from CoinGecko may not be the same as the conversion rate used by the
relayer contract. This commit fixes the gas dropoff slider to use the
conversion rate from the relayer contract instead. A side effect of this
change is that the native gas amount is not displayed in the UI until
the conversion rate is fetched from the relayer contract, after the
debounced debounced value from the slider changes.

Addresses #1076