Fix transfer completion bug in manual transfer claims
When claiming a manual transfer, the useTrackTransfer hook sometimes marked the transfer as complete before the async function triggered by the claim button finished. This allowed users to start a new transfer prematurely, causing the new transfer to be erroneously marked as complete. We should rely solely on the useTrackTransfer hook to set the transfer as complete and handle the redeem transaction. This change prevents the bug and moves the transfer.redeem.start event trigger to before the routes redeem function call.