Show canonical tokens even when wallet is not connected
Wormhole-wrapped Ethereum USDC was not showing with Sui as the source
chain, because it's a Wormhole-wrapped token, and we only display those
if the user has a balance. However, this token is canonical on Sui, so
we want to display it even when the users's wallet isn't connected.

Added an `isCanonical` function that the `TokenList` uses.

Fixes #2690